Renamings in Engine/Serialization
