+++ src/Engine/Snapshots/SnapshotSerialization.h
