+++ src/Engine/Tables/IconFrameTable.cpp
@@ -2,7 +2,7 @@
 
 #include "Engine/Engine.h"
 #include "Engine/Graphics/IRender.h"
-#include "Engine/Serialization/LegacyImages.h"
+#include "Engine/Snapshots/EntitySnapshots.h"
 
 #include "../LOD.h"
