+++ src/Engine/Tables/PlayerFrameTable.cpp
@@ -1,7 +1,7 @@
 #include "PlayerFrameTable.h"
 
 #include "Engine/Engine.h"
-#include "Engine/Serialization/LegacyImages.h"
+#include "Engine/Snapshots/EntitySnapshots.h"
 
 #include "Library/Random/Random.h"
