+++ src/Media/Audio/AudioPlayer.cpp
@@ -15,7 +15,7 @@
 #include "Engine/Objects/Actor.h"
 #include "Engine/Objects/SpriteObject.h"
 #include "Engine/Party.h"
-#include "Engine/Serialization/LegacyImages.h"
+#include "Engine/Snapshots/EntitySnapshots.h"
 
 #include "Media/Audio/OpenALSoundProvider.h"
